requirements.txt
python-dotenv
streamlit
mysql-connector-python
langchain-core
langchain-openai
langchain-groq
langchain-community
flask